Move inlined MMG5_simredmat function in a .h header so symbols are defined in each unit.
