+++ src/common/inlined_functions.h
@@ -305,4 +305,52 @@ inline double MMG5_lenSurfEdg_iso(MMG5_pMesh mesh,MMG5_pSol met,int ip1,int ip2,
   return len;
 }
 
+/**
+ * \param dim matrix size.
+ * \param m matrix array.
+ * \param dm diagonal values array.
+ * \param iv array of inverse coreduction basis.
+ *
+ * Recompose a symmetric matrix from its diagonalization on a simultaneous
+ * reduction basis.
+ * \warning Eigenvectors in Mmg are stored as matrix rows (the first dimension
+ * of the double array spans the number of eigenvectors, the second dimension
+ * spans the number of entries of each eigenvector). So the inverse (left
+ * eigenvectors) is also stored with transposed indices.
+ */
+inline
+void MMG5_simredmat(int8_t dim,double *m,double *dm,double *iv) {
+  int8_t i,j,k,ij;
+
+  /* Storage of a matrix as a one-dimensional array: dim*(dim+1)/2 entries for
+   * a symmetric matrix. */
+  ij = 0;
+
+  /* Loop on matrix rows */
+  for( i = 0; i < dim; i++ ) {
+    /* Loop on the upper-triangular part of the matrix */
+    for( j = i; j < dim; j++ ) {
+      /* Initialize matrix entry */
+      m[ij] = 0.0;
+      /* Compute matrix entry as the recomposition of diagonal values after
+       * projection on the coreduction basis, using the inverse of the
+       * transformation:
+       *
+       * M_{ij} = \sum_{k,l} V^{-1}_{ki} Lambda_{kl} V^{-1}_{lj} =
+       *        = \sum_{k} lambda_{k} V^{-1}_{ki} V^{-1}_{kj}
+       *
+       * Since the inverse of the transformation is the inverse of an
+       * eigenvectors matrix (which is stored in Mmg by columns, and not by
+       * rows), the storage of the inverse matrix is also transposed and the
+       * indices have to be exchanged when implementing the above formula. */
+      for( k = 0; k < dim; k++ ) {
+        m[ij] += dm[k]*iv[i*dim+k]*iv[j*dim+k];
+      }
+      /* Go to the next entry */
+      ++ij;
+    }
+  }
+  assert( ij == (dim+1)*dim/2 );
+}
+
 #endif
