+++ src/common/mettools.c
@@ -33,6 +33,7 @@
  */
 
 #include "mmgcommon.h"
+#include "inlined_functions.h"
 
 /**
  * \param mesh pointer toward the mesh structure.
